Implement new weighted easing.
Take inspiration from https://github.com/kchapelier/hexagrid-relaxing
and implement relaxation based on weighted centroids.

Also refactors the way we handle edges.. the whole NORTH,EAST,SOUTH,WEST
deal was a bad idea for the kind of funky topology we have. Also the
node in the center has 6 connections if it isn't merged.

So we change it to dynamically handle up to 6 six edges per node which
makes our graph very slightly bigger.

This also solves an literal edge case where a quad as two edges on the
outmost edge and our way to store that fact collapses. We just avoid
merging the outmost tris as *first* tri at all which avoids that case.